requirements.txt

crawl4ai>=0.7.2
openai>=1.0.0
google-generativeai>=0.8.0
